# openapi.json
{"openapi":"3.1.0","info":{"title":"Rent Platform API","description":"API for the Rent Platform application","version":"1.0.0"},"paths":{"/users/register":{"post":{"tags":["users"],"summary":"Register","description":"Register a new user","operationId":"register_users_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/verify":{"post":{"tags":["users"],"summary":"Verify","description":"Verify user email with code","operationId":"verify_users_verify_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserVerify"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/resend-verification":{"post":{"tags":["users"],"summary":"Resend Verification","description":"Resend verification code","operationId":"resend_verification_users_resend_verification_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendVerification"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/login":{"post":{"tags":["users"],"summary":"Login","description":"Login user","operationId":"login_users_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"tags":["users"],"summary":"Get Me","description":"Get current user info","operationId":"get_me_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/forgot-password":{"post":{"tags":["users"],"summary":"Forgot Password","description":"Request password reset code","operationId":"forgot_password_users_forgot_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPassword"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/reset-password":{"post":{"tags":["users"],"summary":"Reset Password","description":"Reset password with code","operationId":"reset_password_users_reset_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPassword"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/change-password":{"post":{"tags":["users"],"summary":"Change Password","description":"Change current user's password","operationId":"change_password_users_change_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePassword"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/make-admin/{email}":{"get":{"tags":["users","admin_utils"],"summary":"Make Admin","description":"Find a user by email and promote them to admin.\nUSE WITH CAUTION.","operationId":"make_admin_users_make_admin__email__get","parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string","title":"Email"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/redeem-code":{"post":{"tags":["users","user_actions"],"summary":"Redeem Promo Code","description":"Redeem a promotion code to add balance to the user's account.","operationId":"redeem_promo_code_users_redeem_code_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RedeemCodeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/add-balance":{"post":{"tags":["users","admin_utils"],"summary":"Add Balance","description":"Add balance to a user's account. Admin access required.","operationId":"add_balance_users_add_balance_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAddBalance"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/products/list":{"get":{"tags":["products"],"summary":"List Products","description":"Get list of all products (public endpoint)","operationId":"list_products_products_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProductResponse"},"type":"array","title":"Response List Products Products List Get"}}}}}}},"/products/add":{"post":{"tags":["products"],"summary":"Add Product","description":"Add a new product (admin only)","operationId":"add_product_products_add_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/products/{product_id}":{"put":{"tags":["products","admin_actions"],"summary":"Update Product","description":"Update a product's details (admin only)","operationId":"update_product_products__product_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"integer","title":"Product Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/products/calc-otp":{"get":{"tags":["products"],"summary":"Calc Otp","description":"Calculate OTP from secret key","operationId":"calc_otp_products_calc_otp_get","parameters":[{"name":"secret","in":"query","required":true,"schema":{"type":"string","title":"Secret"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CalcOtpResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders/buy/{product_id}":{"post":{"tags":["orders"],"summary":"Buy Product","description":"Buy/rent a product","operationId":"buy_product_orders_buy__product_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"integer","title":"Product Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders/history":{"get":{"tags":["orders"],"summary":"Get History","description":"Get order history for current user","operationId":"get_history_orders_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/OrderResponse"},"type":"array","title":"Response Get History Orders History Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/orders/all":{"get":{"tags":["orders"],"summary":"Get All Orders","description":"Get all orders (admin only)","operationId":"get_all_orders_orders_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/OrderResponse"},"type":"array","title":"Response Get All Orders Orders All Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/orders/renew/{order_id}":{"post":{"tags":["orders","user_actions"],"summary":"Renew Order","description":"Renew an existing order.","operationId":"renew_order_orders_renew__order_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders/admin/assign":{"post":{"tags":["orders","admin_actions"],"summary":"Admin Assign Product","description":"Admin assigns a product to a user for free.","operationId":"admin_assign_product_orders_admin_assign_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_email","in":"query","required":true,"schema":{"type":"string","title":"User Email"}},{"name":"product_id","in":"query","required":true,"schema":{"type":"integer","title":"Product Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders/admin/revoke/{order_id}":{"delete":{"tags":["orders","admin_actions"],"summary":"Admin Revoke Order","description":"Admin revokes (deletes) a user's order.","operationId":"admin_revoke_order_orders_admin_revoke__order_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders/admin/{order_id}":{"put":{"tags":["orders","admin_actions"],"summary":"Admin Update Order","description":"Admin updates an order's details (e.g., expiration date).","operationId":"admin_update_order_orders_admin__order_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","title":"Order Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/users":{"get":{"tags":["admin"],"summary":"Get All Users With Details","description":"Get a list of all users with their balances and order histories.\nAdmin access required.","operationId":"get_all_users_with_details_admin_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AdminUserDetailResponse"},"type":"array","title":"Response Get All Users With Details Admin Users Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/admin/users/balance":{"put":{"tags":["admin"],"summary":"Set User Balance","description":"Set a user's balance to a specific amount. Admin access required.","operationId":"set_user_balance_admin_users_balance_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAddBalance"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/admin/promo-codes":{"get":{"tags":["admin"],"summary":"Get All Promo Codes","description":"Get a list of all created promotion codes. Admin access required.","operationId":"get_all_promo_codes_admin_promo_codes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/PromoCodeResponse"},"type":"array","title":"Response Get All Promo Codes Admin Promo Codes Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["admin"],"summary":"Create Promo Code","description":"Create a new promotion code. Admin access required.","operationId":"create_promo_code_admin_promo_codes_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromoCodeCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromoCodeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/admin/promo-codes/{code_id}":{"delete":{"tags":["admin"],"summary":"Deactivate Promo Code","description":"Deactivate (soft delete) a promotion code. Admin access required.","operationId":"deactivate_promo_code_admin_promo_codes__code_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"code_id","in":"path","required":true,"schema":{"type":"integer","title":"Code Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AdminUserDetailResponse":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","title":"Email"},"role":{"type":"string","title":"Role"},"balance":{"type":"number","title":"Balance"},"orders":{"items":{"$ref":"#/components/schemas/OrderResponse"},"type":"array","title":"Orders","default":[]}},"type":"object","required":["id","email","role","balance"],"title":"AdminUserDetailResponse"},"AuthResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user":{"$ref":"#/components/schemas/UserResponse"}},"type":"object","required":["access_token","user"],"title":"AuthResponse"},"CalcOtpResponse":{"properties":{"otp":{"type":"string","title":"Otp"}},"type":"object","required":["otp"],"title":"CalcOtpResponse"},"ChangePassword":{"properties":{"old_password":{"type":"string","title":"Old Password"},"new_password":{"type":"string","title":"New Password"}},"type":"object","required":["old_password","new_password"],"title":"ChangePassword"},"ForgotPassword":{"properties":{"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["email"],"title":"ForgotPassword"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MessageResponse":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"MessageResponse"},"OrderResponse":{"properties":{"id":{"type":"integer","title":"Id"},"product_name":{"type":"string","title":"Product Name"},"price":{"type":"number","title":"Price"},"account_info":{"type":"string","title":"Account Info"},"password_info":{"type":"string","title":"Password Info"},"otp_info":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Otp Info"},"purchase_time":{"type":"string","format":"date-time","title":"Purchase Time"},"expires_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Expires At"},"is_expired":{"type":"boolean","title":"Is Expired","default":false},"user_email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Email"},"user_balance_after":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"User Balance After"}},"type":"object","required":["id","product_name","price","account_info","password_info","otp_info","purchase_time","expires_at"],"title":"OrderResponse"},"OrderUpdate":{"properties":{"expires_at":{"type":"string","format":"date-time","title":"Expires At"}},"type":"object","required":["expires_at"],"title":"OrderUpdate"},"ProductCreate":{"properties":{"name":{"type":"string","title":"Name"},"price":{"type":"number","title":"Price"},"quantity":{"type":"integer","title":"Quantity"},"duration":{"type":"string","title":"Duration"},"account":{"type":"string","title":"Account"},"password":{"type":"string","title":"Password"},"otp_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Otp Secret"}},"type":"object","required":["name","price","quantity","duration","account","password"],"title":"ProductCreate"},"ProductResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"price":{"type":"number","title":"Price"},"quantity":{"type":"integer","title":"Quantity"},"duration":{"type":"string","title":"Duration"}},"type":"object","required":["id","name","price","quantity","duration"],"title":"ProductResponse"},"ProductUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price"},"quantity":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Quantity"},"duration":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Duration"},"account_info":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Account Info"},"password_info":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password Info"},"otp_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Otp Secret"}},"type":"object","title":"ProductUpdate"},"PromoCodeCreate":{"properties":{"code":{"type":"string","title":"Code"},"amount":{"type":"number","title":"Amount"}},"type":"object","required":["code","amount"],"title":"PromoCodeCreate"},"PromoCodeResponse":{"properties":{"id":{"type":"integer","title":"Id"},"code":{"type":"string","title":"Code"},"amount":{"type":"number","title":"Amount"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","code","amount","is_active","created_at"],"title":"PromoCodeResponse"},"RedeemCodeRequest":{"properties":{"code":{"type":"string","title":"Code"}},"type":"object","required":["code"],"title":"RedeemCodeRequest"},"ResendVerification":{"properties":{"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["email"],"title":"ResendVerification"},"ResetPassword":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"reset_code":{"type":"string","title":"Reset Code"},"new_password":{"type":"string","title":"New Password"}},"type":"object","required":["email","reset_code","new_password"],"title":"ResetPassword"},"UserAddBalance":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"amount":{"type":"number","title":"Amount"}},"type":"object","required":["email","amount"],"title":"UserAddBalance"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin"},"UserRegister":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserRegister"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","title":"Email"},"role":{"type":"string","title":"Role"},"balance":{"type":"number","title":"Balance"}},"type":"object","required":["id","email","role","balance"],"title":"UserResponse"},"UserVerify":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"code":{"type":"string","title":"Code"}},"type":"object","required":["email","code"],"title":"UserVerify"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/users/login"}}}}}}